Fix: http/https thing in $this->route
As now the request is passed by traefik instead of nginx, it seems that
the proxy headers changes in a way and $this->request is still being
http instead of https. Changing this option it seems that it trust all
the proxies and magically http is being converted to https, then it
matches $this->route with $this->request and the Backpack filters in
Laravel works again.